+++ CHANGELOG.md
@@ -1,2 +1,7 @@
+## 1.1.0
+- Added a way open the Poll Creator menu by sending a phrase in chat (defaulted to `/poll`, configurable).
+- Added config options for automatically closing the Poll Creator menu after pressing the start/end poll button.
+- Added Poll Creator menu scaling (configurable).
+
 ## 1.0.0
-Initial release
+Initial release.